+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.3] - Unreleased
 
+### Fixed
+
+- `Cycles::EndOf` to have the correct behavior
+
 ## [0.1.2] - 2024-08-09
 
 ### Added